Changed conversion of input from int to long long (64-bit) to support large input values (contribute by kallenosf). Added checkValidResultDigits functions to exit the program on unsupported result values.

+++ basecalculator.h
@@ -4,6 +4,6 @@
 
 basenumber baseCalculator(basenumber number1, int basesLength);
 char validMathOperator(void);
-void switchPositions(int *number1, int *number2);
+void switchPositions(long long *number1, long long *number2);
 
 #endif
